+++ js/data.js
@@ -1,0 +1,23 @@
+const MIN_LAT = 35.65000;
+const MAX_LAT = 35.70000;
+
+const MIN_LNG = 139.70000;
+const MAX_LNG = 139.80000;
+
+const OFFER_TITLE = ['Купить Квартиру', 'Купить комнату', 'Купить и сразу продать', 'Продам гараж'];
+
+const OFFER_TYPE = ['palace', 'flat', 'house', 'bungalow', 'hotel'];
+
+const OFFER_TIMES = ['12:00', '13:00', '14:00'];
+
+
+const OFFER_FEATURES = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'];
+
+const OFFER_DESCRIPTION = ['На всю стену — огромное окно, через которое к комнате утром заглядывает солнышко. Посреди помещения — круглый стол коричневого цвета, с изысканной резьбой, украшенный красными гвоздиками в керамическом весе.',
+  'Когда я вхожу в эту комнату, то в глаза сразу бросается большое металлопластиковое окно. Оно выходит во двор. На подоконнике стоят в горшках цветы. Мне нравится зимними вечерами любоваться их зеленью'];
+
+const OFFER_PHOTOS = ['https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg'];
+
+export {MIN_LAT, MAX_LAT, MIN_LNG, MAX_LNG, OFFER_TITLE, OFFER_TYPE, OFFER_TIMES, OFFER_FEATURES, OFFER_DESCRIPTION, OFFER_PHOTOS};
